# openapi.yml
openapi: 3.1.0
info:
  title: MyEstatia AI CRM API
  version: "1.1.0"
  description: |
    API for managing leads, AI-agent communication, companies, properties and system configuration.
servers:
  - url: TODEFINE
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
  /leads:
    get: #DONE
      summary: List all leads
      tags: [Leads]
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"

    post: #DONE
      summary: Create a new lead
      tags: [Leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewLead"
      responses:
        "201":
          description: Lead successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get: #DONE
      summary: Get detailed information for a specific lead
      tags: [Leads]
      parameters:
        - $ref: "#/components/parameters/UUID"
      responses:
        "200":
          description: Lead found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found
...
  /sync/lead/{id}: # synchronize a lead with its related properties (intermediate table: lead_id + property_id) ->
    post: #TODO
      tags: [sync]
      summary: Remove all preferences for the lead and recreate them
      parameters:
        - name: id
          in: path
          required: true
          description: Lead ID
          schema:
            type: string
      responses:
        '200':
          description: Lead found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          description: Lead not found

    delete: #TODO
      tags: [sync]
      summary: Delete all preferences for a lead
      parameters:
        - name: id
          in: path
          required: true
          description: Lead ID
          schema:
            type: string
      responses:
        '200':
          description: Lead found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '404':
          description: Lead not found
...
components:
  parameters:
    UUID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Property:
      type: object
      required: [id, reference, company_id]
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
        company_id:
          type: string
          format: uuid
        title:
          type: string
        address:
          type: string
        rooms:
          type: integer
        price:
          type: number
        description:
          type: string
        created_at:
          type: string
          format: date-time
        image:
          type: string
          example: "https://images.unsplash.com/photo-1613490493576-7fde63acd811?w=800&h=600&fit=crop"
        area:
          type: integer
          example: 180
        bathrooms:
          type: integer
          example: 2
        status:
          type: string
          example: "Disponible"
        source:
          type: string
          example: "Propio"
        compatibleLeadsCount:
          type: integer
          readOnly: true
          example: 8
        isNew:
          type: boolean
          example: true
        zone:
          type: string
          example: "Nueva Andalucía"
...
    Lead:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        rooms:
          type: integer
        budget:
          type: number
        parking:
          type: boolean
        status:
          type: string
        companyId: 
          type: string 
          format: uuid
        propertyId: 
          type: string 
          format: uuid
        assignedAgent:
          type: string
        lastInteraction:
          type: string
          format: date-time
        language:
          type: string
          example: "ES"
        source:
          type: string
          example: "Portal"
        zone:
          type: string
          example: "Nueva Andalucía"
        propertyType:
          type: string
          example: "Villa"
        channel:
          type: string
          example: "WhatsApp"
        suggestedPropertiesCount:
          type: integer
          readOnly: true
          example: 3
        createdAt:
          type: string
          format: date-time
    NewLead:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        rooms:
          type: integer
        budget:
          type: number
        parking:
          type: boolean
        source:
          type: string
...